Update visualization.py
The width and height parameters for Image.new() should be int type.

+++ utils/visualization.py
@@ -17,8 +17,8 @@ def show_img(pixel_array, mode=None):
 
 
 def show_imgs(imgs, width_num=10, height_num=10, x_margin=2, y_margin=2, fpath='/tmp/test.png'):
-    total_width = width_num * IMAGE_SIZE + (width_num-1)*x_margin
-    total_height = height_num * IMAGE_SIZE + (height_num-1)*y_margin
+    total_width = int(width_num * IMAGE_SIZE + (width_num-1)*x_margin)
+    total_height = int(height_num * IMAGE_SIZE + (height_num-1)*y_margin)
 
     new_im = Image.new('RGB', (total_width, total_height), (255,255,255))
